+++ tests/unit/test_clusterer_api/test_sklearn_clustering.py
@@ -22,3 +22,14 @@ def test_cluster(test_df: Dataset):
     )
     assert f"_cluster_.{vector_field}.{alias}" in test_df.schema
     assert len(clusterer.list_closest_to_center()) > 0
+
+
+def test_dbscan(test_client: Client, test_df: Dataset):
+    from sklearn.cluster import DBSCAN
+
+    ALIAS = "dbscan"
+
+    model = DBSCAN()
+    clusterer = test_client.ClusterOps(alias=ALIAS, model=model)
+    clusterer.fit_predict_update(test_df, ["sample_3_vector_"])
+    assert any([x for x in test_df.schema if ALIAS in x])