+++ docsrc/source/changelog.rst
@@ -11,8 +11,12 @@ v0.33.7
 - Add fix and test for new cluster aggregate
 - Add document mocking utility
 - Add integration for cluster reporting
+<<<<<<< HEAD
+- Fix bug for sklearn clustering
+=======
 - Add segment tracking with option to turn off
 - Add print statement after inserting
+>>>>>>> development
 
 v0.33.6
 ---------